Fixed event sync object

+++ include/chrono.hpp
@@ -4,7 +4,7 @@
 #include <ntddk.h>
 
 namespace ktl::chrono {
-using duration_t = uint64_t;
+using duration_t = uint32_t;
 
 LARGE_INTEGER to_native_100ns_tics(duration_t period) noexcept {
   LARGE_INTEGER native_period;
